requirements.txt

fastapi
uvicorn
python-multipart
pillow
imagehash
sqlalchemy
jinja2
aiofiles
